+++ packages/babel-preset-expo/src/index.ts
@@ -6,10 +6,9 @@ import { expoRouterBabelPlugin } from './expo-router-plugin';
 import { lazyImports } from './lazyImports';
 
 type BabelPresetExpoPlatformOptions = {
+  /** @deprecated Set `jsxRuntime: 'classic'` to disable automatic JSX handling.  */
   useTransformReactJSXExperimental?: boolean;
   disableImportExportTransform?: boolean;
-  // Defaults to undefined, set to something truthy to disable `@babel/plugin-transform-react-jsx-self` and `@babel/plugin-transform-react-jsx-source`.
-  withDevTools?: boolean;
   // Defaults to undefined, set to `true` to disable `@babel/plugin-transform-flow-strip-types`
   disableFlowStripTypesTransform?: boolean;
   // Defaults to undefined, set to `false` to disable `@babel/plugin-transform-runtime`
@@ -73,28 +72,10 @@ function babelPresetExpo(api: ConfigAPI, options: BabelPresetExpoOptions = {}): 
     ]);
   }
 
-  // Set true to disable `@babel/plugin-transform-react-jsx`
-  // we override this logic outside of the metro preset so we can add support for
-  // React 17 automatic JSX transformations.
-  // If the logic for `useTransformReactJSXExperimental` ever changes in `metro-react-native-babel-preset`
-  // then this block should be updated to reflect those changes.
-  if (!platformOptions.useTransformReactJSXExperimental) {
-    extraPlugins.push([
-      isDev
-        ? require('@babel/plugin-transform-react-jsx-development')
-        : require('@babel/plugin-transform-react-jsx'),
-      {
-        // Defaults to `automatic`, pass in `classic` to disable auto JSX transformations.
-        runtime: (options && options.jsxRuntime) || 'automatic',
-        ...(options &&
-          options.jsxRuntime !== 'classic' && {
-            importSource: (options && options.jsxImportSource) || 'react',
-          }),
-      },
-    ]);
-    // Purposefully not adding the deprecated packages:
-    // `@babel/plugin-transform-react-jsx-self` and `@babel/plugin-transform-react-jsx-source`
-    // back to the preset.
+  if (platformOptions.useTransformReactJSXExperimental != null) {
+    throw new Error(
+      `babel-preset-expo: The option 'useTransformReactJSXExperimental' has been removed in favor of { jsxRuntime: 'classic' }.`
+    );
   }
 
   const aliasPlugin = getAliasPlugin();
@@ -124,8 +105,6 @@ function babelPresetExpo(api: ConfigAPI, options: BabelPresetExpoOptions = {}): 
         // Reference: https://github.com/expo/expo/pull/4685#discussion_r307143920
         require('metro-react-native-babel-preset'),
         {
-          // Defaults to undefined, set to something truthy to disable `@babel/plugin-transform-react-jsx-self` and `@babel/plugin-transform-react-jsx-source`.
-          withDevTools: platformOptions.withDevTools ?? isDev,
           // Defaults to undefined, set to `true` to disable `@babel/plugin-transform-flow-strip-types`
           disableFlowStripTypesTransform: platformOptions.disableFlowStripTypesTransform,
           // Defaults to undefined, set to `false` to disable `@babel/plugin-transform-runtime`
@@ -140,6 +119,9 @@ function babelPresetExpo(api: ConfigAPI, options: BabelPresetExpoOptions = {}): 
           // TransformError App.js: /path/to/App.js: Duplicate __self prop found. You are most likely using the deprecated transform-react-jsx-self Babel plugin.
           // Both __source and __self are automatically set when using the automatic jsxRuntime. Please remove transform-react-jsx-source and transform-react-jsx-self from your Babel config.
           useTransformReactJSXExperimental: true,
+          // This will never be used regardless because `useTransformReactJSXExperimental` is set to `true`.
+          // https://github.com/facebook/react-native/blob/a4a8695cec640e5cf12be36a0c871115fbce9c87/packages/react-native-babel-preset/src/configs/main.js#L151
+          withDevTools: false,
 
           disableImportExportTransform: platformOptions.disableImportExportTransform,
           lazyImportExportTransform:
@@ -154,6 +136,34 @@ function babelPresetExpo(api: ConfigAPI, options: BabelPresetExpoOptions = {}): 
               : // Pass the option directly to `metro-react-native-babel-preset`, which in turn
                 // passes it to `babel-plugin-transform-modules-commonjs`
                 lazyImportsOption,
+        },
+      ],
+
+      // React support with similar options to Metro.
+      // We override this logic outside of the metro preset so we can add support for
+      // React 17 automatic JSX transformations.
+      // The only known issue is the plugin `@babel/plugin-transform-react-display-name` will be run twice,
+      // once in the Metro plugin, and another time here.
+      [
+        require('@babel/preset-react'),
+        {
+          development: isDev,
+
+          // Defaults to `automatic`, pass in `classic` to disable auto JSX transformations.
+          runtime: options?.jsxRuntime || 'automatic',
+          ...(options &&
+            options.jsxRuntime !== 'classic' && {
+              importSource: (options && options.jsxImportSource) || 'react',
+            }),
+
+          // NOTE: Unexposed props:
+
+          // pragma?: string;
+          // pragmaFrag?: string;
+          // pure?: string;
+          // throwIfNamespace?: boolean;
+          // useBuiltIns?: boolean;
+          // useSpread?: boolean;
         },
       ],
     ],